refactor(library): extract definition aggregate helpers for public catalog reuse
